Refactoring for new standards in code base.
