+++ java/arcs/core/crdt/testutil/BUILD
@@ -1,0 +1,23 @@
+load(
+    "//third_party/java/arcs/build_defs:build_defs.bzl",
+    "arcs_kt_library",
+)
+
+licenses(["notice"])
+
+package(default_visibility = ["//java/arcs:allowed-packages"])
+
+arcs_kt_library(
+    name = "generators",
+    testonly = 1,
+    srcs = [
+        "Generators.kt",
+    ],
+    deps = [
+        "//java/arcs/core/common",
+        "//java/arcs/core/crdt",
+        "//java/arcs/core/data:rawentity",
+        "//java/arcs/core/testutil",
+        "//javatests/arcs/core/testutil",
+    ],
+)
