+++ java/arcs/core/testutil/Fuzzing.kt
@@ -261,7 +261,7 @@ class Value<T>(val value: T) : Generator<T> {
 /**
  * A [Generator] that produces an integer between a min and max.
  */
-class GetInt(
+class IntInRange(
   val s: FuzzingRandom,
   val min: Int,
   val max: Int
