+++ javatests/arcs/core/crdt/CrdtEntityFuzzTest.kt
@@ -1,12 +1,15 @@
 package arcs.core.crdt
 
+import arcs.core.crdt.testutil.CrdtEntityGenerator
+import arcs.core.crdt.testutil.RawEntityGenerator
+import arcs.core.crdt.testutil.ReferencableGenerator
+import arcs.core.crdt.testutil.SingleActorVersionMapGenerator
 import arcs.core.testutil.ChooseFromList
-import arcs.core.testutil.GetInt
+import arcs.core.testutil.IntInRange
 import arcs.core.testutil.MapOf
 import arcs.core.testutil.SetOf
 import arcs.core.testutil.Value
 import arcs.core.testutil.runFuzzTest
-import com.google.common.truth.Truth
 import org.junit.Test
 
 class CrdtEntityFuzzTest() {
@@ -14,16 +17,16 @@ class CrdtEntityFuzzTest() {
    * Test that a randomly generated [CrdtEntity] merged with itself produces an empty [CrdtChange].
    */
   @Test
-  fun crdtEntity_merge_invariant() = runFuzzTest {
-    val versionMap = VersionMapGenerator(
+  fun mergeWithSelf_producesNoChanges() = runFuzzTest {
+    val versionMap = SingleActorVersionMapGenerator(
       ChooseFromList(it, listOf("me", "foo", "bar", "fooBar")),
-      GetInt(it, 1, 20)
+      IntInRange(it, 1, 20)
     )
 
     val singletons = MapOf(
       ChooseFromList(it, listOf("koala", "kangaroo", "penguin", "emu")),
       ReferencableGenerator(ChooseFromList(it, listOf("animal", "mammal", "reptile", "shelled"))),
-      GetInt(it, 0, 3)
+      IntInRange(it, 0, 3)
     )
 
     val collections = MapOf(
@@ -32,12 +35,12 @@ class CrdtEntityFuzzTest() {
         ReferencableGenerator(
           ChooseFromList(
             it,
-            listOf("animals, mammals, reptiles", "shelleds", "pets", "birds", "fish")
+            listOf("animals", "mammals", "reptiles", "shelled", "pets", "birds", "fish")
           )
         ),
-        GetInt(it, 1, 3)
+        IntInRange(it, 1, 3)
       ),
-      GetInt(it, 0, 4)
+      IntInRange(it, 0, 4)
     )
 
     val rawEntity = RawEntityGenerator(
@@ -48,13 +51,11 @@ class CrdtEntityFuzzTest() {
       Value(100)
     )
 
-    val entity1 = CrdtEntityGenerator(
+    val entity = CrdtEntityGenerator(
       versionMap,
       rawEntity
     )()
 
-    val changes = entity1.merge(entity1.data)
-    Truth.assertThat(changes.modelChange.isEmpty()).isTrue()
-    Truth.assertThat(changes.otherChange.isEmpty()).isTrue()
+    invariant_mergeWithSelf_producesNoChanges(entity)
   }
 }
